# openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beyond Queue",
    "description": "PostgreSQL-native message queue with SQS-compatible and REST APIs.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1"
  },
  "paths": {
...
    "/v1/queues/{name}/messages": {
      "get": {
        "tags": [
          "messages"
        ],
        "operationId": "receive_messages",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Queue name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "max",
            "in": "query",
            "description": "Maximum number of messages to return. Default: 1.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "wait",
            "in": "query",
            "description": "Long-poll wait time in seconds. Default: 0 (no wait).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "vt",
            "in": "query",
            "description": "Visibility timeout in seconds. Default: 30.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "fifo",
            "in": "query",
            "description": "Return messages in FIFO order within a group. Default: false.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "messages"
        ],
        "operationId": "send_messages",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Queue name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "async_commit",
            "in": "query",
            "description": "Skip WAL fsync on commit. Improves throughput at the cost of durability:\nmessages can be lost on a PostgreSQL crash. Default: false (durable).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "messages"
        ],
        "operationId": "delete_batch",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Queue name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/queues/{name}/messages/{id}": {
      "delete": {
        "tags": [
          "messages"
        ],
        "operationId": "delete_message",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Queue name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Message ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Message deleted"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "messages"
        ],
        "operationId": "change_visibility",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Queue name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Message ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeVisibilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeVisibilityResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChangeVisibilityRequest": {
        "type": "object",
        "required": [
          "vt"
        ],
        "properties": {
          "vt": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ChangeVisibilityResponse": {
        "type": "object",
        "required": [
          "id",
          "visible_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "visible_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DeleteBatchRequest": {
        "type": "object",
        "required": [
          "ids"
        ],
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "DeletedResponse": {
        "type": "object",
        "required": [
          "deleted"
        ],
        "properties": {
          "deleted": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "MessageResponse": {
        "type": "object",
        "required": [
          "id",
          "read_count",
          "enqueued_at",
          "visible_at",
          "message"
        ],
        "properties": {
          "enqueued_at": {
            "type": "string",
            "format": "date-time"
          },
          "headers": {},
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "message": {},
          "read_count": {
            "type": "integer",
            "format": "int32"
          },
          "visible_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SendBody": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/SendRequest"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SendRequest"
            }
          }
        ]
      },
      "SendRequest": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "delay": {
            "type": "integer",
            "format": "int32"
          },
          "group_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "headers": {},
          "message": {}
        }
      },
      "SendResponse": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "id"
            ],
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "ids"
            ],
            "properties": {
              "ids": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        ]
      },
...
    }
  },
...
}